extract JTS and Testing modules
