+++ .travis.yml
@@ -6,7 +6,7 @@ cache:
     - $HOME/.sbt/boot/
 
 script:
-  - sbt +geohexJVM/test geohexJS/fullOptJS +geohexJVM/package
+  - sbt +test geohexJS/fullOptJS +geohexJVM/package +geohexJts/package +geohexTesting/package
   # Tricks to avoid unnecessary cache updates
   - find $HOME/.sbt -name "*.lock" | xargs rm
   - find $HOME/.ivy2 -name "ivydata-*.properties" | xargs rm
